feat(team): add user team membership check
Adds a check in the TeamListView to see if the current user is a member of any of the teams.
If the user is a member of a team, the team's ID is added to the context.
This allows for conditional rendering of team-related content based on user membership.
Also adds a check in the TeamDetailView to see if the current user is a member of the team.
If the user is a member, a flag is set in the context.
This allows for conditional rendering of team-related content based on user membership.
Finally, adds functionality to update a team with workers and projects.
The code is updated to allow the user to select workers and projects to associate with a team, and these selections are saved when the team is updated.
This makes it easier to manage team membership and project assignments.